Overhaul the test architecture
There was a bug where I was only running the first set of tests in the
main function, which was tiresome. Fixing it let me trim a lot of
redundant code, however, so that's good, and this'll make it easier to
have the "main executable" in this library (if there is one) be an
example server.