I planned to use this file for something, but found a better way.
Trim it.

Specifically, I'd planned on making a top-level server to combine all
the test servers, so that I could run an example server in the exe for
this library. But, I wound up having to do that anyway, as I was only
running one of the TestGroups, even though my main function called all
of them.